Plan 6/legacy/reproducibility_package/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "caphs-feedback-scheduler"
version = "0.1.0"
description = "Minimal reproducibility package for CAPHS heterogeneous scheduling"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "pyyaml>=6.0",
  "networkx>=3.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
